Feature/populate Palma Nova bars

+++ DataStructureService/models/crawls_bars.py
@@ -3,5 +3,5 @@ from database import db
 
 crawls_bars = db.Table('crawls_bars',
     db.Column('crawl_id', db.Integer, db.ForeignKey('crawl.id'), primary_key=True),
-    db.Column('bar_id', db.Integer, db.ForeignKey('bar.id'), primary_key=True)
+    db.Column('bar_id', db.BigInteger, db.ForeignKey('bar.id'), primary_key=True)
 )
